+++ src/core/default-config.js
@@ -2,12 +2,14 @@ const defaultConfig = {
     directives: {
         fieldValidationErrorClassNames: 'field-validation-error-class',
         fieldValidationErrorMessage: 'field-validation-error-message',
+        formStateClassNames: 'form-state-class',
+        formStateMessage: 'form-state-message',
         formSubmissionErrorClassNames: 'form-submission-error-class',
         formSubmissionErrorMessage: 'form-submission-error-message',
         formSubmissionSuccessClassNames: 'form-submission-success-class',
         formSubmissionSuccessMessage: 'form-submission-success-message',
         formValidationErrorClassNames: 'form-validation-error-class',
-        formValidationErrorMessage: 'form-validation-error-message'
+        formValidationErrorMessage: 'form-validation-error-message',
     },
     classNames: {
         fieldValidationError: 'is-error is-field-validation-error',