+++ src/core/directives-handler-map.js
@@ -1,8 +1,10 @@
 import setFieldsValidationErrorMessage from '../directives/set-fields-validation-error-message';
+import setFormStateMessage from '../directives/set-form-state-message';
 import setFormSubmissionErrorMessage from '../directives/set-form-submission-error-message';
 import setFormSubmissionSuccessMessage from '../directives/set-form-submission-success-message';
 import setFormValidationErrorMessage from '../directives/set-form-validation-error-message';
 import toggleFieldsValidationErrorClassNames from '../directives/toggle-fields-validation-error-classnames';
+import toggleFormStateClassNames from '../directives/toggle-form-state-classnames';
 import toggleFormSubmissionErrorClassNames from '../directives/toggle-form-submission-error-classnames';
 import toggleFormSubmissionSuccessClassNames from '../directives/toggle-form-submission-success-classnames';
 import toggleFormValidationErrorClassNames from '../directives/toggle-form-validation-error-classnames';
@@ -10,6 +12,8 @@ import toggleFormValidationErrorClassNames from '../directives/toggle-form-valid
 const directivesHandlerMap = {
     fieldsValidationErrorClassNames: setFieldsValidationErrorMessage,
     fieldsValidationErrorMessage: toggleFieldsValidationErrorClassNames,
+    formStateClassNames: toggleFormStateClassNames,
+    formStateMessage: setFormStateMessage,
     formSubmissionErrorClassNames: toggleFormSubmissionErrorClassNames,
     formSubmissionErrorMessage: setFormSubmissionErrorMessage,
     formSubmissionSuccessClassNames: toggleFormSubmissionSuccessClassNames,